Seperating chart component from the stats page
