+++ src/styles/Home.module.css
@@ -214,3 +214,12 @@
   font-family: var(--font-family-base);
   font-weight: var(--font-weight-base);
 }
+
+.chart {
+  width: 100%;
+  height: 300px;
+  margin: 0 auto;
+  padding: 0;
+  border: 0;
+  display: block;
+}
